Completed assignment

+++ Binomial.java
@@ -21,8 +21,8 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	 
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		int [][] memo = new int [n + 1][k + 1];
+	public static long binomial(int n, int k) {
+		long [][] memo = new int [n + 1][k + 1];
 		 for(int i = 0; i < memo.length; i++) {
 			for(int j = 0; j < memo[0].length; j++) {
 				memo[i][j] = -1;
@@ -32,7 +32,7 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	}
 
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
